# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trust Corp Bank API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
    put:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/kyc:
    post:
      operationId: submitKyc
      tags: [users]
      summary: Submit KYC information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycInput"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [credit, debit, transfer, topup, exchange]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
  /transactions/{txId}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single transaction
      parameters:
        - name: txId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
  /transactions/send:
    post:
      operationId: sendMoney
      tags: [transactions]
      summary: Send money to another account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendInput"
      responses:
        "201":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient funds or invalid input
  /transactions/topup:
    post:
      operationId: topUpAccount
      tags: [transactions]
      summary: Top up an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopUpInput"
      responses:
        "201":
          description: Top-up completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, email, fullName, kycStatus, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        kycStatus:
          type: string
          enum: [pending, submitted, approved, rejected]
        address:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        address:
          type: string
        dateOfBirth:
          type: string
    KycInput:
      type: object
      required: [fullName, dateOfBirth, address, idType, idNumber]
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        address:
          type: string
        idType:
          type: string
          enum: [passport, drivers_license, national_id]
        idNumber:
          type: string
        phone:
          type: string
...
    Transaction:
      type: object
      required: [id, accountId, type, amount, currency, status, description, createdAt]
      properties:
        id:
          type: integer
        accountId:
          type: integer
        type:
          type: string
          enum: [credit, debit, transfer, topup, exchange]
        amount:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, completed, failed, reversed]
        description:
          type: string
        reference:
          type: ["string", "null"]
        recipientName:
          type: ["string", "null"]
        recipientAccount:
          type: ["string", "null"]
        senderName:
          type: ["string", "null"]
        balanceAfter:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TransactionList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    SendInput:
      type: object
      required: [fromAccountId, amount, currency, description]
      properties:
        fromAccountId:
          type: integer
        beneficiaryId:
          type: ["integer", "null"]
        recipientAccount:
          type: ["string", "null"]
        recipientName:
          type: ["string", "null"]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
    TopUpInput:
      type: object
      required: [accountId, amount, currency]
      properties:
        accountId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        source:
          type: string
          default: external
...
    Notification:
      type: object
      required: [id, userId, title, message, type, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [transaction, security, kyc, system, promotion]
        read:
          type: boolean
        createdAt:
          type: string
...