# lib/api-zod/src/generated/types/topUpInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trust Corp Bank API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TopUpInput {
  accountId: number;
  amount: number;
  currency: string;
  source?: string;
}